SA-970: Move Tooltip from signals to shared component
