+++ src/pages/signals/EngagementRateByCohortPage.js
@@ -12,7 +12,7 @@ import EngagementRateByCohortActions from './components/actionContent/Engagement
 import Divider from './components/Divider';
 import Page from './components/SignalsPage';
 import Tabs from './components/engagement/Tabs';
-import TooltipMetric from './components/charts/tooltip/TooltipMetric';
+import TooltipMetric from 'src/components/charts/TooltipMetric';
 import withDetails from './containers/withDetails';
 import withDateSelection from './containers/withDateSelection';
 import { ENGAGEMENT_RECENCY_COHORTS } from './constants/info';
