+++ src/pages/signals/dashboards/components/HealthScoreChart/HealthScoreChart.js
@@ -4,7 +4,7 @@ import { getCaretProps, getDoD } from 'src/helpers/signals';
 import { selectCurrentHealthScoreDashboard } from 'src/selectors/signals';
 import { Panel } from '@sparkpost/matchbox';
 import BarChart from '../../../components/charts/barchart/BarChart';
-import TooltipMetric from '../../../components/charts/tooltip/TooltipMetric';
+import TooltipMetric from 'src/components/charts/TooltipMetric';
 import { PanelLoading } from 'src/components';
 import Callout from 'src/components/callout';
 import MetricDisplay from '../MetricDisplay/MetricDisplay';
